# Faster internal group_keys implementation