This appeared to be taking unnecessary time, presumably because of the infamous `row.names` handling in R, so I worked out it. 

``` r
library(dplyr)

df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100))) %>% group_by(g)

bench::mark(
  group_keys(df), 
  group_data(df)[-2], 
  
  check = FALSE # not exactly the same because `group_keys()` also removes the .drop attribute
)
#> # A tibble: 2 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 group_keys(df)       4.37µs   4.97µs   171603.    27.6KB     17.2
#> 2 group_data(df)[-2] 345.28µs  413.9µs     2482.    41.3KB     14.6
```